MAINT: Highlight アクションをすべて一つのコンポーネントにまとめた
